Switch to AJAX read status
Changed 'mark as' button to be a button that does nothing rather than a link to the update action.
Created the api endpoint for changing the read status upon request.

Still need to remove the read status updating from the standard update action so it lives only in the api's links controller.